Remove initialize env | remove price-check-consumer in favor of SQS trigger event processing
