+++ index.ts
@@ -1,19 +1,39 @@
 import 'module-alias/register';
-import initializeEnv from '@/utils/initializeEnv';
 import bootstrap from './src/bootstrap';
-import PriceCheckConsumer from '@/resources/price-check-consumer/price-check.consumer';
+import { SQSEvent } from 'aws-lambda';
+import DiscountManager from '@/resources/discount-manager/discount-manager';
+import { removeS3KeySuffix, sleep } from '@/resources/resource.utils';
+import { SqsMsgBody } from './src/types';
 
-export const handler = async (): Promise<void> => {
-    console.log('Initializing environment');
-    return initializeEnv().then(async () => {
-        bootstrap();
+let manager: DiscountManager;
+let frequency: number;
 
-        console.log('Initializing polling');
-        const polling = [
-            new PriceCheckConsumer()
-        ].map(consumer => consumer.startPolling());
-        return Promise.all(polling).then(() => {
-            console.log('Polling complete for all consumers');
-        });
-    });
+export const handler = async (event: SQSEvent): Promise<void> => {
+    console.log(`Recieved event: ${JSON.stringify(event)}`);
+    bootstrap();
+    manager = new DiscountManager();
+    frequency = +(process.env.price_check_consumer_frequency ?? 1);
+
+    for(let record of event.Records) {
+        try {
+            const message: SqsMsgBody = JSON.parse(record.body);
+            processSqsEvent(message);
+        } catch (err: any) {
+            const cik = record.body;
+            await manager.intiateDiscountCheck(cik);
+        }
+        await sleep(1000 * frequency);
+    };
+
+    console.log('Processing complete!');
 };
+
+async function processSqsEvent(event: SqsMsgBody): Promise<void> {
+    for (let record of event.Records) {
+        if (record.s3 && record.s3.object && record.s3.object.key) {
+            const cik = removeS3KeySuffix(record.s3.object.key);
+            console.log(`In price check consumer, processing: ${cik}`);
+            return manager.intiateDiscountCheck(cik);
+        }
+    }
+}
