--- main.ts
@@ -1,3 +1,0 @@
-import { handler } from "./index";
-
-handler();
