+++ package.json
@@ -6,8 +6,7 @@
   "scripts": {
     "test": "ts-mocha --paths -p ./tsconfig.json test/**/*.spec.ts --exit",
     "coverage": "nyc --reporter=lcov --reporter=text-summary npm run test",
-    "start": "ts-node main.ts",
-    "dev": "tsc-watch --onSuccess \"ts-node main.ts\"",
+    "dev": "tsc-watch",
     "build": "tsc"
   },
   "extends": "@istanbuljs/nyc-config-typescript",
@@ -55,7 +54,6 @@
     "node-fetch": "^2.6.9",
     "nyc": "^15.1.0",
     "random-words": "^1.3.0",
-    "sqs-consumer": "^7.2.2",
     "typescript": "^5.2.2"
   },
   "_moduleAliases": {
