+++ src/resources/discount-manager/discount-manager.ts
