+++ src/resources/discount-manager/discount-manager.typings.ts
