+++ src/resources/discount-manager/discount-manager.utils.ts
